# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "更好的MAS游戏社区版",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/accounts/update": {
      "post": {
        "tags": [
          "Community"
        ],
        "summary": "Update Account",
        "operationId": "updateAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutBase"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts/delete": {
      "post": {
        "tags": [
          "Community"
        ],
        "summary": "Delete Account",
        "operationId": "deleteAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountIdIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutBase"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts/reorder": {
      "post": {
        "tags": [
          "Community"
        ],
        "summary": "Reorder Accounts",
        "operationId": "reorderAccounts",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountReorderIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutBase"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings": {
      "get": {
        "tags": [
          "Community"
        ],
        "summary": "Get Settings",
        "operationId": "getSettings",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsOut"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Community"
        ],
        "summary": "Update Settings",
        "operationId": "updateSettings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutBase"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/activity": {
      "post": {
        "tags": [
          "Community"
        ],
        "summary": "Query Activity",
        "operationId": "queryActivity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivityQueryIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityActivityOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/login/{provider}/qr/create": {
      "post": {
        "tags": [
          "Community"
        ],
        "summary": "Create Qr",
        "operationId": "createQr",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "miyoushe",
                "skland"
              ],
              "type": "string",
              "title": "Provider"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QrCreateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/login/{provider}/qr/check": {
      "post": {
        "tags": [
          "Community"
        ],
        "summary": "Check Qr",
        "operationId": "checkQr",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "miyoushe",
                "skland"
              ],
              "type": "string",
              "title": "Provider"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QrCheckIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QrCheckOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/login/{provider}/qr/save": {
      "post": {
        "tags": [
          "Community"
        ],
        "summary": "Save Qr",
        "operationId": "saveQr",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "miyoushe",
                "skland"
              ],
              "type": "string",
              "title": "Provider"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QrSaveIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutBase"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/login/taygedo": {
      "post": {
        "tags": [
          "Community"
        ],
        "summary": "Login Taygedo",
        "operationId": "loginTaygedo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaygedoLoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutBase"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountIdIn": {
        "properties": {
          "accountId": {
            "type": "string",
            "title": "Accountid"
          }
        },
        "type": "object",
        "required": [
          "accountId"
        ],
        "title": "AccountIdIn"
      },
...
      "AccountReorderIn": {
        "properties": {
          "order": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Order"
          }
        },
        "type": "object",
        "required": [
          "order"
        ],
        "title": "AccountReorderIn"
      },
      "AccountUpdateIn": {
        "properties": {
          "accountId": {
            "type": "string",
            "title": "Accountid"
          },
          "data": {
            "$ref": "#/components/schemas/GameSignAccountGroupConfig"
          }
        },
        "type": "object",
        "required": [
          "accountId",
          "data"
        ],
        "title": "AccountUpdateIn"
      },
...
      "ActivityQueryIn": {
        "properties": {
          "accountIds": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accountids"
          }
        },
        "type": "object",
        "title": "ActivityQueryIn"
      },
      "CommunityActivityOut": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码",
            "default": 200
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "操作状态",
            "default": "success"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "操作消息",
            "default": "操作成功"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/CommunityActivitySnapshotOut"
            },
            "type": "array",
            "title": "Data",
            "description": "按账号和游戏拆分的活动快照"
          }
        },
        "type": "object",
        "title": "CommunityActivityOut",
        "description": "游戏社区日常活动查询响应。"
      },
...
      "CommunityActivitySnapshotOut": {
        "properties": {
          "account": {
            "type": "string",
            "title": "Account",
            "description": "账号组名称"
          },
          "accountUid": {
            "type": "string",
            "title": "Accountuid",
            "description": "账号组 UUID"
          },
          "game": {
            "type": "string",
            "title": "Game",
            "description": "游戏名称"
          },
          "platform": {
            "type": "string",
            "title": "Platform",
            "description": "社区平台名称"
          },
          "status": {
            "type": "string",
            "enum": [
              "success",
              "empty",
              "limited",
              "unavailable",
              "failed"
            ],
            "title": "Status",
            "description": "活动查询状态"
          },
          "completed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed",
            "description": "已完成数量"
          },
          "target": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target",
            "description": "目标数量"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/CommunityActivityTaskOut"
            },
            "type": "array",
            "title": "Tasks",
            "description": "每日任务"
          },
          "resources": {
            "items": {
              "$ref": "#/components/schemas/CommunityActivityResourceOut"
            },
            "type": "array",
            "title": "Resources",
            "description": "可用资源"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "description": "失败或受限原因",
            "default": ""
          },
          "updatedAt": {
            "type": "string",
            "title": "Updatedat",
            "description": "查询时间",
            "default": ""
          },
          "roleName": {
            "type": "string",
            "title": "Rolename",
            "description": "角色名称",
            "default": ""
          },
          "roleUid": {
            "type": "string",
            "title": "Roleuid",
            "description": "角色 UID",
            "default": ""
          },
          "server": {
            "type": "string",
            "title": "Server",
            "description": "角色区服",
            "default": ""
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "已确认的数据来源路径",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "account",
          "accountUid",
          "game",
          "platform",
          "status"
        ],
        "title": "CommunityActivitySnapshotOut",
        "description": "单个游戏角色的日常活动快照。"
      },
      "CommunityActivityTaskOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "任务名称"
          },
          "completed": {
            "type": "integer",
            "title": "Completed",
            "description": "已完成数量"
          },
          "target": {
            "type": "integer",
            "title": "Target",
            "description": "目标数量"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "任务状态"
          },
          "period": {
            "type": "string",
            "title": "Period",
            "description": "任务周期",
            "default": "daily"
          }
        },
        "type": "object",
        "required": [
          "name",
          "completed",
          "target",
          "status"
        ],
        "title": "CommunityActivityTaskOut",
        "description": "日常活动中的单项任务。"
      },
...
      "GameSignAccountGroupConfig": {
        "properties": {
          "Name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "账号组名称"
          },
          "Enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled",
            "description": "是否启用"
          },
          "MiyousheToken": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Miyoushetoken",
            "description": "米游社登录凭证"
          },
          "MiyousheDeviceId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Miyoushedeviceid",
            "description": "米游社安卓设备 ID，仅用于绝区零便笺"
          },
          "MiyousheDeviceFp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Miyoushedevicefp",
            "description": "米游社安卓设备指纹，仅用于绝区零便笺"
          },
          "CloudGenshinToken": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cloudgenshintoken",
            "description": "云原神 combo token"
          },
          "KuroToken": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kurotoken",
            "description": "库街区登录凭证"
          },
          "SklandToken": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sklandtoken",
            "description": "森空岛登录凭证"
          },
          "TaygedoToken": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taygedotoken",
            "description": "塔吉多及云异环登录凭证"
          },
          "LastSignDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastsigndate",
            "description": "账号组上次签到日期"
          }
        },
        "type": "object",
        "title": "GameSignAccountGroupConfig",
        "description": "游戏社区账号组配置"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OutBase": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码",
            "default": 200
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "操作状态",
            "default": "success"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "操作消息",
            "default": "操作成功"
          }
        },
        "type": "object",
        "title": "OutBase"
      },
      "QrCheckIn": {
        "properties": {
          "ticket": {
            "type": "string",
            "title": "Ticket"
          },
          "device": {
            "type": "string",
            "title": "Device"
          }
        },
        "type": "object",
        "required": [
          "ticket",
          "device"
        ],
        "title": "QrCheckIn"
      },
      "QrCheckOut": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码",
            "default": 200
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "操作状态",
            "default": "success"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "操作消息",
            "default": "操作成功"
          },
          "cookies_str": {
            "type": "string",
            "title": "Cookies Str",
            "default": ""
          },
          "scan_code": {
            "type": "string",
            "title": "Scan Code",
            "default": ""
          }
        },
        "type": "object",
        "title": "QrCheckOut"
      },
      "QrCreateOut": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码",
            "default": 200
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "操作状态",
            "default": "success"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "操作消息",
            "default": "操作成功"
          },
          "ticket": {
            "type": "string",
            "title": "Ticket",
            "default": ""
          },
          "qr_url": {
            "type": "string",
            "title": "Qr Url",
            "default": ""
          },
          "device": {
            "type": "string",
            "title": "Device",
            "default": ""
          }
        },
        "type": "object",
        "title": "QrCreateOut"
      },
      "QrSaveIn": {
        "properties": {
          "account_uid": {
            "type": "string",
            "title": "Account Uid"
          },
          "cookie": {
            "type": "string",
            "title": "Cookie",
            "default": ""
          },
          "scan_code": {
            "type": "string",
            "title": "Scan Code",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "account_uid"
        ],
        "title": "QrSaveIn"
      },
...
      "SettingsData": {
        "properties": {
          "Enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "ActivityEnabled": {
            "type": "boolean",
            "title": "Activityenabled",
            "default": true
          },
          "MiyousheBbsEnabled": {
            "type": "boolean",
            "title": "Miyoushebbsenabled",
            "default": true
          },
          "RunOnStartup": {
            "type": "boolean",
            "title": "Runonstartup",
            "default": false
          },
          "ScheduledRun": {
            "type": "boolean",
            "title": "Scheduledrun",
            "default": false
          },
          "ScheduledTime": {
            "type": "string",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "title": "Scheduledtime",
            "default": "08:00"
          },
          "LowPerformanceMode": {
            "type": "boolean",
            "title": "Lowperformancemode",
            "default": false
          },
          "Theme": {
            "type": "string",
            "enum": [
              "light",
              "dark",
              "system"
            ],
            "title": "Theme",
            "default": "system"
          },
          "Proxy": {
            "type": "string",
            "title": "Proxy",
            "default": ""
          },
          "CaptchaMode": {
            "type": "string",
            "enum": [
              "manual",
              "local",
              "local_yunma"
            ],
            "title": "Captchamode",
            "default": "local"
          },
          "YunmaToken": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Yunmatoken"
          },
          "MasBaseUrl": {
            "type": "string",
            "maxLength": 200,
            "title": "Masbaseurl",
            "default": "http://127.0.0.1:36163"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SettingsData"
      },
      "SettingsOut": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码",
            "default": 200
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "操作状态",
            "default": "success"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "操作消息",
            "default": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/SettingsData"
          },
          "yunmaConfigured": {
            "type": "boolean",
            "title": "Yunmaconfigured",
            "default": false
          },
          "localConnections": {
            "type": "boolean",
            "title": "Localconnections",
            "description": "是否能连接后端主机的 MAS 与代理；纯 Workers 为 false",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SettingsOut"
      },
...
      "TaygedoLoginIn": {
        "properties": {
          "accountId": {
            "type": "string",
            "title": "Accountid"
          },
          "phone": {
            "type": "string",
            "minLength": 1,
            "title": "Phone"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "format": "password",
            "title": "Password",
            "writeOnly": true
          }
        },
        "type": "object",
        "required": [
          "accountId",
          "phone",
          "password"
        ],
        "title": "TaygedoLoginIn"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}